refactor: seperate runner and simulators
