+++ pyproject.toml
@@ -31,7 +31,7 @@ ignore_missing_imports = true
 
 [tool.coverage.run]
 omit = ["*/__init__.py", 
-        "*/_OnDemandRunner.py",
+        "*/_OnDemandSimulator.py",
         "*/_qpuRunner.py",
         ]
