+++ src/orquestra/integrations/braket/simulator/__init__.py
@@ -2,8 +2,15 @@
 # © Copyright 2021-2022 Zapata Computing Inc.
 ################################################################################
 
+try:
+    from ._LocalSimulator import BraketLocalSimulator
+except ModuleNotFoundError:
+    pass
 
 try:
-    from ._qpuRunner import BraketQPURunner, get_QPU_names
+    from ._OnDemandSimulator import (
+        BraketOnDemandSimulator,
+        get_on_demand_simulator_names,
+    )
 except ModuleNotFoundError:
     pass
